requirements.txt
aiomqtt==1.2.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncio-mqtt==0.16.2
black==23.7.0
certifi==2023.7.22
click==8.1.7
constants==0.6.0
dnspython==2.4.2
docopt==0.6.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.101.1
fastapi-mqtt==1.1.1
gmqtt==0.6.12
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.2.0
mqtt-client==1.6.1
mypy-extensions==1.0.0
orjson==3.9.5
packaging==23.1
paho-mqtt==1.6.1
pathspec==0.11.2
platformdirs==3.10.0
pydantic==1.10.12
pydantic-extra-types==2.0.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pymongo==4.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.0
sniffio==1.3.0
starlette==0.27.0
terminaltables==3.1.10
tomli==2.0.1
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
